Implement Add Account
This Commit:
- Creates the add account layout
- Implements the add account input fields
- Implements some basic navigation
- Adds logic layer and interface to logic layer
- Adds some more functionality to logic and persistence layer

+++ app/src/main/java/com/example/financemanager/objects/Account.java
@@ -6,11 +6,11 @@ public class Account {
     private String name;
     private String bank;
     private String type;
-    private int balance;
+    private double balance;
 
     private boolean isExpanded;
 
-    public Account(int accountID, String name, String bank, String type, int balance) {
+    public Account(int accountID, String name, String bank, String type, double balance) {
         this.accountID = accountID;
         this.name = name;
         this.bank = bank;
@@ -51,11 +51,11 @@ public class Account {
         this.type = type;
     }
 
-    public int getBalance() {
+    public double getBalance() {
         return balance;
     }
 
-    public void setBalance(int balance) {
+    public void setBalance(double balance) {
         this.balance = balance;
     }
